requirements.txt
openai>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langgraph>=0.0.40
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
typing-extensions>=4.5.0
pydantic>=2.0.0